replace in place the helpers file
